Update README.md
TMS FNC Gantt Chart v1.2.2.1 Full Source

With the TMS FNC Gantt Chart you will get the project management tool to provide you with a visual of what needs to be done and when. The Gantt Chart is a component for tracking your project schedules. Where you can define if tasks rely on others, the progress that is made and the current status of your tasks.

[TMS FNC Gantt Chart v1.2.2.1 Full Source](https://developer.team/delphi/34917-tms-fnc-gantt-chart-v1221-full-source.html)